+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - rich module for pretty printing
 - Parser class to work with ttp
 - Unit tests for Parser class
+- cisco-ios ttp template
 
 ### Fixed